feat: combine flask and dash app
